+++ src/main/php/io/streams/compress/Bzip2.class.php
@@ -16,15 +16,19 @@ class Bzip2 implements Algorithm {
   /** Returns the algorithm's common file extension, including a leading "." */
   public function extension(): string { return '.bz2'; }
 
+  /** Maps fastest, default and strongest levels */
+  public function level(int $select): int {
+    static $levels= [Compression::FASTEST => 1, Compression::DEFAULT => 4, Compression::STRONGEST => 9];
+    return $levels[$select] ?? $select;
+  }
+
   /** Opens an input stream for reading */
   public function open(InputStream $in): InputStream {
     return new Bzip2InputStream($in);
   }
 
   /** Opens an output stream for writing */
-  public function create(OutputStream $out, int $method= Compression::DEFAULT): OutputStream {
-    static $levels= [Compression::FASTEST => 1, Compression::DEFAULT => 4, Compression::STRONGEST => 9];
-
-    return new Bzip2OutputStream($out, $levels[$method]);
+  public function create(OutputStream $out, int $level= Compression::DEFAULT): OutputStream {
+    return new Bzip2OutputStream($out, $this->level($level));
   }
 }