+++ src/main/php/io/streams/compress/Brotli.class.php
@@ -2,7 +2,7 @@
 
 use io\streams\{InputStream, OutputStream, Compression};
 
-class Brotli implements Algorithm {
+class Brotli extends Algorithm {
 
   /** Returns whether this algorithm is supported in the current setup */
   public function supported(): bool { return extension_loaded('brotli'); }