+++ src/main/php/io/streams/compress/Algorithm.class.php
@@ -1,8 +1,9 @@
 <?php namespace io\streams\compress;
 
 use io\streams\{Compression, InputStream, OutputStream};
+use lang\Value;
 
-abstract class Algorithm {
+abstract class Algorithm implements Value {
 
   /** Returns whether this algorithm is supported in the current setup */
   public abstract function supported(): bool;
@@ -24,4 +25,30 @@ abstract class Algorithm {
 
   /** Opens an output stream for writing */
   public abstract function create(OutputStream $out, int $level= Compression::DEFAULT): OutputStream;
+
+  /** @return string */
+  public function hashCode() { return crc32($this->name); }
+
+  /** @return string */
+  public function toString() {
+    return sprintf(
+      '%s(token: %s, extension: %s, supported: %s, levels: %d..%d)',
+      nameof($this),
+      $this->token(),
+      $this->extension() ?: '(none)',
+      $this->supported() ? 'true' : 'false',
+      $this->level(Compression::FASTEST),
+      $this->level(Compression::STRONGEST)
+    );
+  }
+
+  /**
+   * Compare this algorithm to a given value.
+   *
+   * @param  var $value
+   * @return int
+   */
+  public function compareTo($value) {
+    return $value instanceof self ? $this->name() <=> $value->name() : 1;
+  }
 }