+++ ChangeLog.md
@@ -3,6 +3,9 @@ Compression streams ChangeLog
 
 ## ?.?.? / ????-??-??
 
+## 1.0.0 / 2023-05-16
+
+* Fixed GZIP header reading - @thekid
 * Merged PR #3: Migrate to new testing library - @thekid
 
 ## 1.0.0 / 2022-02-26